+++ lib/modules/_storage/Memory.ts
@@ -1,3 +1,4 @@
+import {isFunction} from '@valkyriestudios/utils/function';
 import {isObject} from '@valkyriestudios/utils/object';
 import {isIntGt} from '@valkyriestudios/utils/number';
 import {isNeString} from '@valkyriestudios/utils/string';
@@ -12,20 +13,42 @@ export class MemoryStore <T extends TriFrostStoreValue = TriFrostStoreValue> imp
 
     #store = new Map<string, {value:T; expires?:number}>();
 
+    /* Garbage collection interval */
     #gc:ReturnType<typeof setInterval>|null = null;
 
-    constructor (opts?: {gc_interval?: number; gc_filter?: GCFilter<T>}) {
-        const interval = opts?.gc_interval;
-        const filter = opts?.gc_filter ?? ((_key, _v, now, _exp) => isIntGt(_exp, 0) && _exp <= now);
+    /* Used for lru (least-recently-used) tracking */
+    #lru:Set<string> = new Set();
 
-        if (interval) {
+    /* Set to the max amount of items allowed in our store if configured */
+    #lruMax:number|null = null;
+
+    constructor (opts?: {
+        gc_interval?: number;
+        gc_filter?: GCFilter<T>;
+        max_items?: number;
+    }) {
+        /* Configure garbage collection interval */
+        const filter = isFunction(opts?.gc_filter)
+            ? opts.gc_filter
+            : (_key:string, _v:T, _now:number, _exp?:number) => isIntGt(_exp, 0) && _exp <= _now;
+        if (isIntGt(opts?.gc_interval, 0)) {
             this.#gc = setInterval(() => {
                 const now = Date.now();
                 for (const [key, entry] of this.#store.entries()) {
-                    if (filter(key, entry.value, now, entry.expires !== undefined ? entry.expires : null)) this.#store.delete(key);
+                    if (filter(key, entry.value, now, entry.expires)) {
+                        this.#store.delete(key);
+                        if (this.isLRU) this.#lru.delete(key);
+                    }
                 }
-            }, interval);
+            }, opts.gc_interval);
         }
+
+        /* Set max usage value if max entries is provided */
+        if (isIntGt(opts?.max_items, 0)) this.#lruMax = opts.max_items;
+    }
+
+    private get isLRU () {
+        return this.#lruMax !== null;
     }
 
     async get (key: string): Promise<T|null> {
@@ -36,10 +59,18 @@ export class MemoryStore <T extends TriFrostStoreValue = TriFrostStoreValue> imp
 
         if (isIntGt(val.expires, 0) && Date.now() > val.expires) {
             this.#store.delete(key);
+            /* Remove from LRU */
+            if (this.isLRU) this.#lru.delete(key);
             return null;
-        }
+        } else {
+            /* Mark as most recently used in LRU */
+            if (this.isLRU) {
+                this.#lru.delete(key);
+                this.#lru.add(key);
+            }
 
-        return val.value as T;
+            return val.value as T;
+        }
     }
 
     async set (
@@ -50,6 +81,22 @@ export class MemoryStore <T extends TriFrostStoreValue = TriFrostStoreValue> imp
         if (!isNeString(key)) throw new Error('TriFrostMemoryStore@set: Invalid key');
         if (!isObject(value) && !Array.isArray(value)) throw new Error('TriFrostMemoryStore@set: Invalid value');
 
+        /* If configured as an LRU, update and see if we need to evict any keys */
+        if (this.isLRU) {
+            /* Mark as most recently used in LRU */
+            this.#lru.delete(key);
+            this.#lru.add(key);
+
+            /* Evict if above size */
+            if (this.#lru.size > this.#lruMax!) {
+                const to_evict = this.#lru.values().next().value;
+                if (to_evict) {
+                    this.#store.delete(to_evict);
+                    this.#lru.delete(to_evict);
+                }
+            }
+        }
+
         const expires = isIntGt(opts?.ttl, 0)
             ? Date.now() + (opts.ttl * 1000)
             : undefined;
@@ -59,6 +106,9 @@ export class MemoryStore <T extends TriFrostStoreValue = TriFrostStoreValue> imp
     async delete (key: string): Promise<void> {
         if (!isNeString(key)) throw new Error('TriFrostMemoryStore@delete: Invalid key');
         this.#store.delete(key);
+        
+        /* Remove from lru */
+        if (this.isLRU) this.#lru.delete(key);
     }
 
     async stop () {